fix init file
